Removed all pam/quantize function from imageio_png and moved them in a new file
